refactor: split risk engine into focused modules
